requirements.txt
altair==5.4.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
distro==1.9.0
exceptiongroup==1.2.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.9.0
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.0.2
openai==1.51.0
packaging==24.1
pandas==2.2.3
phidata==2.4.39
pillow==10.4.0
protobuf==5.28.2
pyarrow==17.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
rich==13.9.1
rpds-py==0.20.0
serpapi==0.1.5
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
streamlit==1.39.0
tenacity==9.0.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.20.0
toml==0.10.2
tomli==2.0.2
tornado==6.4.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
youtube-transcript-api==0.6.2
